Separate string namespace
