+++ py-polars/tests/unit/namespaces/test_string.py
@@ -1,8 +1,199 @@
-from __future__ import annotations
+from datetime import date, datetime
 
 import pytest
 
 import polars as pl
+from polars.testing import assert_series_equal
+
+
+def test_str_slice() -> None:
+    df = pl.DataFrame({"a": ["foobar", "barfoo"]})
+    assert df["a"].str.slice(-3).to_list() == ["bar", "foo"]
+    assert df.select([pl.col("a").str.slice(2, 4)])["a"].to_list() == ["obar", "rfoo"]
+
+
+def test_str_concat() -> None:
+    s = pl.Series(["1", None, "2"])
+    result = s.str.concat()
+    expected = pl.Series(["1-null-2"])
+    assert_series_equal(result, expected)
+
+
+def test_str_lengths() -> None:
+    s = pl.Series(["Café", None, "345", "東京"])
+    expected = pl.Series([5, None, 3, 6], dtype=pl.UInt32)
+    assert_series_equal(s.str.lengths(), expected)
+
+
+def test_str_n_chars() -> None:
+    s = pl.Series(["Café", None, "345", "東京"])
+    expected = pl.Series([4, None, 3, 2], dtype=pl.UInt32)
+    assert_series_equal(s.str.n_chars(), expected)
+
+
+def test_str_contains() -> None:
+    s = pl.Series(["messi", "ronaldo", "ibrahimovic"])
+    expected = pl.Series([True, False, False])
+    assert_series_equal(s.str.contains("mes"), expected)
+
+
+def test_str_encode() -> None:
+    s = pl.Series(["foo", "bar", None])
+    hex_encoded = pl.Series(["666f6f", "626172", None])
+    base64_encoded = pl.Series(["Zm9v", "YmFy", None])
+    assert_series_equal(s.str.encode("hex"), hex_encoded)
+    assert_series_equal(s.str.encode("base64"), base64_encoded)
+    with pytest.raises(ValueError):
+        s.str.encode("utf8")  # type: ignore[arg-type]
+
+
+def test_str_decode() -> None:
+    hex_encoded = pl.Series(["666f6f", "626172", None])
+    base64_encoded = pl.Series(["Zm9v", "YmFy", None])
+    expected = pl.Series([b"foo", b"bar", None])
+
+    assert_series_equal(hex_encoded.str.decode("hex"), expected)
+    assert_series_equal(base64_encoded.str.decode("base64"), expected)
+
+
+def test_str_decode_exception() -> None:
+    s = pl.Series(["not a valid", "626172", None])
+    with pytest.raises(Exception):
+        s.str.decode(encoding="hex")
+    with pytest.raises(Exception):
+        s.str.decode(encoding="base64")
+    with pytest.raises(ValueError):
+        s.str.decode("utf8")  # type: ignore[arg-type]
+
+
+def test_str_replace_str_replace_all() -> None:
+    s = pl.Series(["hello", "world", "test", "root"])
+    expected = pl.Series(["hell0", "w0rld", "test", "r0ot"])
+    assert_series_equal(s.str.replace("o", "0"), expected)
+
+    expected = pl.Series(["hell0", "w0rld", "test", "r00t"])
+    assert_series_equal(s.str.replace_all("o", "0"), expected)
+
+
+def test_str_to_lowercase() -> None:
+    s = pl.Series(["Hello", "WORLD"])
+    expected = pl.Series(["hello", "world"])
+    assert_series_equal(s.str.to_lowercase(), expected)
+
+
+def test_str_to_uppercase() -> None:
+    s = pl.Series(["Hello", "WORLD"])
+    expected = pl.Series(["HELLO", "WORLD"])
+    assert_series_equal(s.str.to_uppercase(), expected)
+
+
+def test_str_parse_int() -> None:
+    bin = pl.Series(["110", "101", "010"])
+    assert_series_equal(bin.str.parse_int(2), pl.Series([6, 5, 2]).cast(pl.Int32))
+
+    hex = pl.Series(["fa1e", "ff00", "cafe"])
+    assert_series_equal(
+        hex.str.parse_int(16), pl.Series([64030, 65280, 51966]).cast(pl.Int32)
+    )
+
+
+def test_str_strip() -> None:
+    s = pl.Series([" hello ", "world\t "])
+    expected = pl.Series(["hello", "world"])
+    assert_series_equal(s.str.strip(), expected)
+
+    expected = pl.Series(["hello", "worl"])
+    assert_series_equal(s.str.strip().str.strip("d"), expected)
+
+    expected = pl.Series(["ell", "rld\t"])
+    assert_series_equal(s.str.strip(" hwo"), expected)
+
+
+def test_str_lstrip() -> None:
+    s = pl.Series([" hello ", "\t world"])
+    expected = pl.Series(["hello ", "world"])
+    assert_series_equal(s.str.lstrip(), expected)
+
+    expected = pl.Series(["ello ", "world"])
+    assert_series_equal(s.str.lstrip().str.lstrip("h"), expected)
+
+    expected = pl.Series(["ello ", "\t world"])
+    assert_series_equal(s.str.lstrip("hw "), expected)
+
+
+def test_str_rstrip() -> None:
+    s = pl.Series([" hello ", "world\t "])
+    expected = pl.Series([" hello", "world"])
+    assert_series_equal(s.str.rstrip(), expected)
+
+    expected = pl.Series([" hell", "world"])
+    assert_series_equal(s.str.rstrip().str.rstrip("o"), expected)
+
+    expected = pl.Series([" he", "wor"])
+    assert_series_equal(s.str.rstrip("odl \t"), expected)
+
+
+def test_str_strip_whitespace() -> None:
+    s = pl.Series("a", ["trailing  ", "  leading", "  both  "])
+
+    expected = pl.Series("a", ["trailing", "  leading", "  both"])
+    assert_series_equal(s.str.rstrip(), expected)
+
+    expected = pl.Series("a", ["trailing  ", "leading", "both  "])
+    assert_series_equal(s.str.lstrip(), expected)
+
+    expected = pl.Series("a", ["trailing", "leading", "both"])
+    assert_series_equal(s.str.strip(), expected)
+
+
+def test_str_strptime() -> None:
+    s = pl.Series(["2020-01-01", "2020-02-02"])
+    expected = pl.Series([date(2020, 1, 1), date(2020, 2, 2)])
+    assert_series_equal(s.str.strptime(pl.Date, "%Y-%m-%d"), expected)
+
+    s = pl.Series(["2020-01-01 00:00:00", "2020-02-02 03:20:10"])
+    expected = pl.Series(
+        [datetime(2020, 1, 1, 0, 0, 0), datetime(2020, 2, 2, 3, 20, 10)]
+    )
+    assert_series_equal(s.str.strptime(pl.Datetime, "%Y-%m-%d %H:%M:%S"), expected)
+
+    s = pl.Series(["00:00:00", "03:20:10"])
+    expected = pl.Series([0, 12010000000000], dtype=pl.Time)
+    assert_series_equal(s.str.strptime(pl.Time, "%H:%M:%S"), expected)
+
+
+def test_str_split() -> None:
+    a = pl.Series("a", ["a, b", "a", "ab,c,de"])
+    for out in [a.str.split(","), pl.select(pl.lit(a).str.split(",")).to_series()]:
+        assert out[0].to_list() == ["a", " b"]
+        assert out[1].to_list() == ["a"]
+        assert out[2].to_list() == ["ab", "c", "de"]
+
+    for out in [
+        a.str.split(",", inclusive=True),
+        pl.select(pl.lit(a).str.split(",", inclusive=True)).to_series(),
+    ]:
+        assert out[0].to_list() == ["a,", " b"]
+        assert out[1].to_list() == ["a"]
+        assert out[2].to_list() == ["ab,", "c,", "de"]
+
+
+def test_jsonpath_single() -> None:
+    s = pl.Series(['{"a":"1"}', None, '{"a":2}', '{"a":2.1}', '{"a":true}'])
+    expected = pl.Series(["1", None, "2", "2.1", "true"])
+    assert_series_equal(s.str.json_path_match("$.a"), expected)
+
+
+def test_extract_regex() -> None:
+    s = pl.Series(
+        [
+            "http://vote.com/ballon_dor?candidate=messi&ref=polars",
+            "http://vote.com/ballon_dor?candidat=jorginho&ref=polars",
+            "http://vote.com/ballon_dor?candidate=ronaldo&ref=polars",
+        ]
+    )
+    expected = pl.Series(["messi", None, "ronaldo"])
+    assert_series_equal(s.str.extract(r"candidate=(\w+)", 1), expected)
 
 
 def test_extract_binary() -> None:
@@ -91,12 +282,6 @@ def test_contains_expr() -> None:
         df.select(
             pl.col("text").str.contains(pl.col("pattern"), literal=False, strict=True)
         )
-
-
-def test_null_comparisons() -> None:
-    s = pl.Series("s", [None, "str", "a"])
-    assert (s.shift() == s).null_count() == 0
-    assert (s.shift() != s).null_count() == 0
 
 
 def test_replace() -> None:
@@ -243,21 +428,6 @@ def test_ljust_and_rjust() -> None:
         "ljust": ["foo       ", "longer_foo", "longest_fooooooo", "hi        "],
         "ljust_len": [10, 10, 16, 10],
     }
-
-
-def test_format_empty_df() -> None:
-    df = pl.DataFrame(
-        [
-            pl.Series("val1", [], dtype=pl.Categorical),
-            pl.Series("val2", [], dtype=pl.Categorical),
-        ]
-    ).select(
-        [
-            pl.format("{}:{}", pl.col("val1"), pl.col("val2")).alias("cat"),
-        ]
-    )
-    assert df.shape == (0, 1)
-    assert df.dtypes == [pl.Utf8]
 
 
 def test_starts_ends_with() -> None:
